Rename Chart component into DataWidget
