+++ src/js/components/DataWidget.jsx
